xy now take user input for the function

+++ inverse_kinematics.py
@@ -31,8 +31,8 @@ def deg(rad):
 	return rad * 180 / math.pi
 
 
-x = int(raw_input("x>"))
-y = int(raw_input("y>"))
+x = float(raw_input("x>"))
+y = float(raw_input("y>"))
 a1, a2 = angles(x, y)
 
 print "x=",x
